+++ releaf-content/lib/releaf/content/node.rb
@@ -51,7 +51,7 @@ module Releaf::Content
       end
 
       def attributes_to_not_copy
-        list = %w[content_id depth id item_position lft rgt slug created_at updated_at]
+        list = %w[content_id depth id item_position lft rgt created_at updated_at]
         list << "locale" if locale_before_type_cast.blank?
         list
       end
@@ -84,6 +84,22 @@ module Releaf::Content
         end
       end
 
+      # Maintain unique slug within parent_id scope.
+      # If slug is not unique add numeric postfix.
+      def maintain_slug
+        postfix = nil
+        total_count = 0
+
+        while self.class.where(parent_id: parent_id, slug: "#{slug}#{postfix}").where("id != ?", id.to_i).exists? do
+          total_count += 1
+          postfix = "-#{total_count}"
+        end
+
+        if postfix
+          self.slug = "#{slug}#{postfix}"
+        end
+      end
+
       # Returns closest existing locale starting from object itself
       # @return [String] locale
       def locale
@@ -100,9 +116,11 @@ module Releaf::Content
       # Check whether object and all its ancestors are active
       # @return [Boolean] returns true if object is available
       def available?
-        # There seams to be bug in Rails 4.0.0, that prevents us from using exists?
-        # exists? will return nil or 1 in this query, instead of true/false (as it should)
-        self_and_ancestors.where(active: false).any? == false
+        self_and_ancestors_array.all?(&:active?)
+      end
+
+      def self_and_ancestors_array
+        preloaded_self_and_ancestors.nil? ? self_and_ancestors.to_a : preloaded_self_and_ancestors
       end
 
       def reasign_slug
@@ -212,7 +230,9 @@ module Releaf::Content
 
       after_save :update_settings_timestamp, unless: :prevent_auto_update_settings_timestamp?
 
-      acts_as_url :name, url_attribute: :slug, scope: :parent_id, :only_when_blank => true
+      acts_as_url :name, url_attribute: :slug, scope: :parent_id, only_when_blank: true
+
+      attr_accessor :preloaded_self_and_ancestors
     end
   end
 end