+++ releaf-core/lib/generators/dummy/templates/builders/admin/banners/show_builder.rb
@@ -1,0 +1,7 @@
+module Admin::Banners
+  class ShowBuilder < Releaf::Builders::ShowBuilder
+    def section_body
+      render "content"
+    end
+  end
+end