+++ releaf-core/lib/generators/dummy/templates/config/routes.rb
@@ -2,6 +2,7 @@ Dummy::Application.routes.draw do
 
   mount_releaf_at '/admin' do
     releaf_resources :books, :authors, :chapters, :publishers
+    releaf_resources :banners, only: [:index, :show]
   end
 
   # SINGLE NODE CLASS CASE: