+++ releaf-core/lib/generators/dummy/templates/controllers/admin/banners_controller.rb
@@ -1,0 +1,5 @@
+class Admin::BannersController < Releaf::ActionController
+  def features
+    [:index, :show]
+  end
+end