+++ releaf-core/lib/generators/dummy/templates/models/banner.rb
@@ -2,4 +2,8 @@ class Banner < ActiveRecord::Base
   dragonfly_accessor :image
   validates_presence_of :url
   belongs_to :banner_group
+
+  def releaf_title
+    url
+  end
 end