+++ spec/rails_helper.rb
@@ -97,6 +97,7 @@ RSpec.configure do |config|
   config.include FactoryGirl::Syntax::Methods
 
   Capybara.javascript_driver = :poltergeist
+  Capybara.server = :webrick
 
   # disable empty translation creation